// wardn/frontend/lib/api/generated/model/mCPRuntimeInstallationControlResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { MCPRuntimeSessionHealthResponse } from './mCPRuntimeSessionHealthResponse';
import type { MCPRuntimeSessionRead } from './mCPRuntimeSessionRead';

export interface MCPRuntimeInstallationControlResponse {
  action: string;
  active: boolean;
  canRedeploy: boolean;
  canRestart: boolean;
  canStart: boolean;
  canStop: boolean;
  health?: MCPRuntimeSessionHealthResponse | null;
  installationId: string;
  runtimeSession?: MCPRuntimeSessionRead | null;
}